add interpolation of time series data prior to image reconstruction

+++ image_reconstruction/reconstruction_utils/beamforming/fft_based_hauptmann_2018.py
@@ -41,7 +41,7 @@ def fft_hauptmann_2d(time_series_data, detection_elements, sampling_rate_hz,
         A 2D numpy array with the following internal array definition: [detectors, time samples]
     detection_elements: dict
         Definition of the transducer elements
-    sampling_rate_hz: int
+    sampling_rate_hz: float
         Data sampling rate in Hz
     field_of_view: np.ndarray
         The target field of view in [xmin, xmax, ymin, ymax, zmin, zmax] in meters
